### 2023-06-02 #5518 @editorialbot
[PRE REVIEW]: COOLEST: the COde-independent Organized LEns STandard
**Submitting author:** @aymgal (<a href="http://orcid.org/0000-0003-2547-9815">Aymeric Galan</a>)
**Repository:** https://github.com/aymgal/COOLEST
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.0.1
