### 2023-06-05 #5518 @aymgal
HI @plaplant , thank you for being the editor and managing the review of our paper. I had a look to the reviewer list, and saw only one person I know with expertise in gravitational lensing (sibirrer) however there may be a conflict of interest here as he is developing lenstronomy, a code for which we have used for checks and have implemented an interface with COOLEST.
