### 2023-06-14 #5518 @aymgal
@plaplant Looking at the people that reviewed the JOSS entry for lenstronomy, I saw that smsharma would be someone with experience with gravitational lensing, given one of his recent works.
