### 2023-06-16 #5567 @editorialbot
[REVIEW]: COOLEST: the COde-independent Organized LEns STandard
**Submitting author:** @aymgal (<a href="http://orcid.org/0000-0003-2547-9815">Aymeric Galan</a>)
**Repository:** https://github.com/aymgal/COOLEST
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
