### 2023-06-23 aymgal/COOLEST#36 @AlexandreAdam
Installation
Please include a small statement in the `README.md` file for the installation method, either with `pip` or directly calling the `setup.py` script. There should be a method to install the package for general usage and one for development / testing.
