### 2023-06-23 aymgal/COOLEST#37 @AlexandreAdam
List of conventions
Please provide the full list of convention in the README.md file, similar to what is done in the documentation.
