### 2023-06-23 aymgal/COOLEST#38 @AlexandreAdam
Hosting the documentation
While individual classes have a neat documentation, I would find it useful to have a dedicated website hosting the documentation. The docs and html for this are already in the joss branch, but I could not find the documentation online. Please provide a link to it in the `README.md` file when it is properly hosted.
