### 2023-06-23 aymgal/COOLEST#39 @AlexandreAdam
Examples in the documentation
Please provide a notebook showcasing the plotting API. A notebook detailing how to reproduce the main figure of the paper should be sufficient.
