### 2023-06-23 #5567 @AlexandreAdam
## Installation

Please include a small statement in the `README.md` file for the installation method, either with `pip` or directly calling the `setup.py` script. There should be a method to install the package for general usage and one for development / testing. I have opened [this issue](https://github.com/aymgal/COOLEST/issues/36) which should be quick to resolve.

## Functionality

The full list of adopted convention is not yet in the `README.md` (joss branch), nor is it in the paper. I have opened [this issue](https://github.com/aymgal/COOLEST/issues/37) for this. Since this is the main output of this project, I strongly suggest that the complete list be provided and described in detail somewhere in the project and in the paper. In itself, this will be very useful for other developer to possibly adjust their convention manually in order to align it with COOLEST.

Also see [this issue](https://github.com/aymgal/COOLEST/issues/35) for bugs in the `/examples` folder.

Notebooks in `/docs/notebooks/` are working fine.

I dug around in the code and found [this notebook](https://github.com/aymgal/COOLEST/blob/joss/coolest/api/test_einstein_rad.ipynb) which describes a different method (compared to lenstronomy) for computing the effective Einstein radii. Speed comparisons with lenstronomy are provided, which I reproduced. This notebook should be moved in the test folder as an explicit test. Now as to the merit of this code, I cannot tell right away if this algorithm is correct. It seems to me like it misses the mark set by an analytical value by a substantial margin. To be fair, so does lenstronomy, yet if this code is used for _reproducibility_ purposes it seems a rather important issue. This last point is subjective, so this is not an acceptance-blocking issue.

## Documentation
While individual classes have a neat documentation, I would find it useful to have a dedicated website hosting the documentation. The docs and html for this are already in the joss branch, but I could not find the documentation online. Please provide a link to it in the `README.md` file when it is properly hosted. See [this issue](https://github.com/aymgal/COOLEST/issues/38).

For the purpose of this review, I hosted locally the documentation and have the following comments

1. I am confused by the author's point in the page `What is a "lensing entity"?` placed in `<host>/template_file/lensing_entities.html.` The concept of a _lensing plane_ is particularly useful, not only to distinguish it from the _source plane_, but also in the context of multiplane lensing. I do agree that the properties of a galaxy, like it's mass profile and light profile, are semantically separate from the lensing plane, but this distinction does not seem relevant for what the authors want. If I understood the context correctly, the authors want a naming convention that attaches physical properties to physical objects while keeping the _lensing formalism_ in the background. I do not disagree with that choice, but the text is rather vague on this and seem to miss the point.

2. It would be great if the examples (notebooks) were more elaborate. For example, a notebook on how to compare models like is showcase in the paper would be great but is missing. I have opened [this issue](https://github.com/aymgal/COOLEST/issues/39).

3. I am completely missing the point described in the page `Modes: mock vs best-fit`. The usage of the `MOCK` mode is not illustrated in the examples, and the `DOC` mode is rather mysterious. In either case, I would want to see how the JSON template structure is modified by these choice (if it is), or what segment of the code is affected by this.

## Tests
There is no documentation or mention in the `README.md` file on how to run tests. I figured the author where using `pytest`, but a small statement should be included in the `README.md` to tell users how to run it (a simple one-liner would be appreciated).

All test pass on my system, although I get a significant amounts of warning related to the usage of deprecated function. The specific warning messages I get from `numpy` look _safe_, so this is not an acceptance-blocker issue.

## Task list for the authors
- [ ] [Issue 35](https://github.com/aymgal/COOLEST/issues/35)
- [ ] [Issue 36](https://github.com/aymgal/COOLEST/issues/36)
- [ ]  [Issue 37](https://github.com/aymgal/COOLEST/issues/37)
- [ ]  [Issue 38](https://github.com/aymgal/COOLEST/issues/38)
- [ ] [Issue 39](https://github.com/aymgal/COOLEST/issues/39)
