### 2023-06-30 #5567 @aymgal
Thank you very much @AlexandreAdam for your thorough review on our work! We will go through the entire list above and github issues, and address your points in the following days/weeks. Once this is done I will summarize this here.

Just a quick remark here: the documentation is hosted [here](https://coolest.readthedocs.io/), as mentioned on the github repo. I will make this even clearer when updating the README based on your suggestions.
