### 2023-07-03 7de034a
Update README with link to documentation and installation instructions
