+++ README.md
@@ -1,3 +1,5 @@
+<img src="docs/_static/coolest_logo.png" width="200" style="transform:rotate(90deg);" alt="Herculens logo" />
+
 # COOLEST: COde-independent Organized LEns STandard
 
 ![License](https://img.shields.io/github/license/aymgal/COOLEST)
@@ -7,37 +9,54 @@
 [![Coverage Status](https://coveralls.io/repos/github/aymgal/COOLEST/badge.svg)](https://coveralls.io/github/aymgal/COOLEST)
 
 
-## A standard for gravitational lens modeling
+## What is COOLEST?
 
-After a lens modeling analysis is published and available as a manuscript, it is often challenging to **(1) reproduce the results** with the same or another similar modeling software, or **(2) start a new analysis** directly based on published results, **(3) reliably and quantitatively compare results** from different analyses. Moreover, there is currently no practical way to share and store lens models within the community.
+COOLEST is a standard that proposes to define a set of conventions to be shared across the strong lensing community, in order to consistently store, share and improve lens modeling analyses. In short, this project provides tools to manipulate lens models as a single, human-readable JSON template file alongside Python routines for visualizing and comparing lens models possibly obtained from different modeling codes.
 
-The **COde-independent Organized LEns STandard (COOLEST)** proposes a solution to the above problems. COOLEST is a standard for **describing, storing and sharing lens models** more easily, independently of the specific modeling techniques and software packages. The main element of COOLEST is a unique JSON hierarchical template file, that stores in a human-readable way the necessary lens model data. Another key element is a suit of visualization tools that automatically produces a series of images, plots and other analysis products from a COOLEST file.
+For more information and the full list of COOLEST features, refer to the [online documentation](https://coolest.readthedocs.io/en/latest/).
 
-Currently, this repository hosts a Python API to __generate__, __manipulate__ and __update__ COOLEST files. The API structure exactly mirrors the hierarchy of the template file. Visualization and analysis classes will soon be added.
+## Components of the standard
 
-Ultimately, the goal is that modeling software packages each have an interface (a simple function) that converts code-specific model choices and parameter values to COOLEST conventions. This way, all lens modeling results---best-fit parameter values, posterior distributions---can be saved effortlessly to this standard, ready to be shared and used for further analyses.
+COOLEST is composed of three distinct building blocks:
+- __A set of [conventions](docs/conventions.md)__: coordinate systems, units and profile definitions implicitly assumed when manipulating lens models stored in the template file;
+- __Template file system__ ([`coolest.template`](coolest/template/)): a Python interface to create, store and manipulate COOLEST template files (JSON format) and linked external files (such as FITS files for data products);
+- __Analysis & plotting__ ([`coolest.api`](coolest/template/)): a Python interface to compute key lensing quantities and generate publication-ready lens model plots.
 
+## Examples
 
+### Compact storage of a lens models
 
-## Adopted conventions
+See the [`examples`](examples) directory to see typical JSON template files describing lens models.
 
-COOLEST defines a set of fixed conventions to define coordinate systems, model parameters such as ellipticities, and how data files are stored/linked to the template. Here is a subset of those conventions:
+### Side-to-side model comparison
 
-- **Units**:
-    - Lengths, radii, angular positions, widths, heights are expressed in arcseconds
-    - Position angles and orientations of elliptical profiles are defined in the interval (-90, +90] degrees
+![Example lens model comparison using the COOLEST plotting and analysis capabilities](joss/coolest_plot_example.png)
 
-- **Coordinate system**:
[…]
