### 2023-07-03 a96f10c
Rotated 90deg the logo
