+++ README.md
@@ -1,4 +1,4 @@
-<img src="docs/_static/coolest_logo.png" width="200" style="transform:rotate(90deg);" alt="Herculens logo" />
+<img src="docs/_static/coolest_logo.png" width="200" alt="Herculens logo" />
 
 # COOLEST: COde-independent Organized LEns STandard
 
