### 2023-07-03 e4a7abc
Fix test install command
