### 2023-07-03 aymgal/COOLEST#36 @aymgal
This [commit](https://github.com/aymgal/COOLEST/commit/7de034a523ce974b964d1536cc57fcdb4631f6a4) should have fixed this issue, by adding installation instructions using pip together with the `setup.py`. Let me know if anything else is still missing.
