### 2023-07-03 aymgal/COOLEST#38 @aymgal
This [commit](https://github.com/aymgal/COOLEST/commit/7de034a523ce974b964d1536cc57fcdb4631f6a4) should have fixed this issue, by adding a link to the documentation hosted on readthedocs.
