### 2023-07-03 016c18d
Clean example directory and gather all notebooks under docs directory
