+++ docs/notebooks/index.md
@@ -11,4 +11,5 @@ maxdepth: 1
 List supported options <01-list_supported_options.ipynb>
 Generate a JSON template file <02-generate_template.ipynb>
 Update a template in 2 ways <03-load_and_update_template.ipynb>
+Plot models side-to-side <04-plot_lens_models.ipynb>
 ```
