+++ examples/notebooks/README.md
@@ -1,3 +1,0 @@
-# Example notebooks
-
-Python notebooks can be found in the documentation directory, under [`docs/notebooks`](/docs/notebooks).
