### 2023-07-03 db54cc5
Add lens image to the doc homepage
