+++ requirements.txt
@@ -1,3 +1,8 @@
 numpy>=1.20.0
+scipy>=1.9.0
 jsonpickle>=2.0.0
 astropy>=4.2.0
+
+# Optional installs
+# matplotlib>=3.7.0  # for plotting
+# lenstronomy>=1.11.0  # for shapelets light profiles
