### 2023-07-03 aymgal/COOLEST#39 @aymgal
There was actually the notebook that generated the paper figure in the `joss` directory. I moved it to the same directory as the other notebooks, [`docs/notebooks`](https://github.com/aymgal/COOLEST/tree/joss/docs/notebooks), which are then rendered in the [online doc](https://coolest.readthedocs.io/en/latest/index.html).
