### 2023-07-03 aymgal/COOLEST#35 @aymgal
These example scripts are actually deprecated... sorry for that @AlexandreAdam. Therefore I removed the example directory so that all example notebooks (and related template files) are grouped into a [single place](https://github.com/aymgal/COOLEST/tree/joss/docs/notebooks). I grouped everything under the documentation directory so there are rendered directly in the documentation website.
Let me know if you face problems running the example notebooks.
