### 2023-07-04 7f06f9c
Add unit tests for Analysis routines
