### 2023-07-04 4211657
Add pytest command to run unit tests and fix deprecated method usage
