### 2023-07-04 8621ea1
Improve the description of lensing entities
