### 2023-07-04 765ac43
Clarify the documenation about the templates modes
