+++ docs/template_file/modes.md
@@ -1,15 +1,17 @@
-# The different template "modes"
+# The different template modes
 
-The JSON template at the core of the standard is not limited to store best-fit parameters of a lens model. Another practical use of COOLEST is to _create_ a lens model and define related parameter values to generate new mock data. The difference with a lens model is that the template file would not contain prior or posterior distribution information, but only (point estimate) values for each model parameter. Another different is that there should not be any data file contained in the COOLEST directory, as it will be generated by the simulator code based on the content of the template.
+The key `"mode"` indicates the purpose of the COOLEST template: 1) storing a previously optimized lens model (default), 2) using as input to generate new mock imaging data, or 3) generate a template filled with additional information and fields. The mode does not change the structure of the file, but only filters out which keys are present in it.
 
-The `'mode'` section of the JSON file informs about the usage of the template. Currently there is three modes:
+The three supported modes are as follows:
 
-- `'MAP'`: if the template is for storing maximum _a posteriori_ (MAP) parameter values, optionally alongside prior distribution information and statistics regarding posterior distributions.
+- `"MAP"` (default): the template stores a lens model that was optimized on some data, and typically contain maximum _a posteriori_ (MAP) parameter values, posterior statistics and optionally prior distribution and parameters that have fixed during modeling. The `"instrument"` and `"observation"` keys should be consistent with the data that was modeled.
 
-- `'MOCK'`: if the template is meant to describe a simulated strong lens system, to be used with a given simulator code.
+- `"MOCK"`: the template stores all necessary information regarding the simulation of a mock lens. The template can then be used an input file to a simulator that shares an interface with COOLEST. The `"instrument"` and `"observation"` keys define the observational properties of the mock (field of view, noise, etc.).
 
-- `'DOC'`: if the template should contains fields that give additional information about each JSON section (only useful for developing documentation or understanding the hierarchical structure of the file).
+- `"DOC"`: the template contains a lot of additional fields that give. This mode does not have explicit purpose other than given more details about the different fields of a COOLEST template (depending on the amount of information, it may become rather cluttered). Typically, the additional fields not present in the other two modes are parameters units, definition intervals, LaTeX strings, etc.
 
 ``` {admonition} Note
-An additional fourth mode is planned, focused on an efficient storage of full posterior sampling chains.
+Only the modes `"MAP"` and `"MOCK"` find obvious applications in a strong lensing analysis. `"DOC"` is mostly meant for development purposes.
 ```
+
+This example notebook shows how to generate templates files. The resulting files can be retrieved in the different modes: [`"MAP"`](../notebooks/template_dir/coolest_template.json), [`"MOCK"`](../notebooks/template_dir/coolest_template_mock.json), [`"DOC"`](../notebooks/template_dir/coolest_template_doc.json).
