### 2023-07-04 #5567 @aymgal
Hi @AlexandreAdam , here are answers to your comments.

- __List of conventions listed in the paper__

If possible, I would prefer to have the list of conventions in the documentation / GitHub only, such that it can be completed and improved along its adoption by researchers and the feedback from the community. Although the current stated conventions should not change, it is still an eventuality and it is probable that other conventions get added to the list in the future. Moreover, as the paper is already over the "standard" size, this would make it significantly longer.

- __Tests of the effective quantities computation__

> I dug around in the code and found this notebook which describes a different method (compared to lenstronomy) for computing the effective Einstein radii. Speed comparisons with lenstronomy are provided, which I reproduced. This notebook should be moved in the test folder as an explicit test. Now as to the merit of this code, I cannot tell right away if this algorithm is correct. It seems to me like it misses the mark set by an analytical value by a substantial margin. To be fair, so does lenstronomy. Yet if this code is used for reproducibility purposes it seems a rather important issue. This last point is subjective, so this is not an acceptance-blocking issue.

We added unit test for various input parameters and grid resolution in test/test. For these tests, we quote a maximum relative error of 5% on Einstein radius, effective kappa slope and effective half-light radius computations (the error being most of the time inferior to 5%).

We note that COOLEST is not meant to _reproduce_ the original modeling results (this would be unrealistic), but to provide tools to analyse lens models uniformly, independently from the original modeling code as much as possible. Therefore, differences in the values of effective quantities compared to original lens modeling codes and theoretical values are expected, but does not impact the goal of the standard.

- __Clarification of the concept of `lensing entities`__

> I am confused by the author's point in the page What is a "lensing entity"? placed in <host>/template_file/lensing_entities.html. The concept of a lensing plane is particularly useful, not only to distinguish it from the source plane, but also in the context of multiplane lensing. I do agree that the properties of a galaxy, like it's mass profile and light profile, are semantically separate from the lensing plane, but this distinction does not seem relevant for what the authors want.

We reformulated the documentation, and you can see the commit diff [here](https://github.com/aymgal/COOLEST/commit/8621ea19deb58908da26456f551d398000a26cf4). Please let me know if this is still unclear. Note that the multiplane aspect is fully retained, as each lensing entity has redshift associated to it.

> If I understood the context correctly, the authors want a naming convention that attaches physical properties to physical objects while keeping the lensing formalism in the background. I do not disagree with that choice, but the text is rather vague on this and seem to miss the point.

This is exactly the goal. It has two main advantages: 1) it is more intuitive and makes it easier to access the information within the template file / through the Python API; 2) on the long run, this will become handy when combining constraints on the properties of physical objects from different analyses (this is relevant in the goal of having a database of lens models, that researchers can query). Hopefully the text is clearer now.

- __Clarification on the template "modes"__

> I am completely missing the point described in the page Modes: mock vs best-fit. The usage of the MOCK mode is not illustrated in the examples, and the DOC mode is rather mysterious. In either case, I would want to see how the JSON template structure is modified by these choice (if it is), or what segment of the code is affected by this.

We clarified the documentation, and you can see the commit diff [here](https://github.com/aymgal/COOLEST/commit/765ac43235576f6b78908702319b1215954de075). We also completed this [example notebook](https://github.com/aymgal/COOLEST/blob/joss/docs/notebooks/02-generate_template.ipynb) such that it also dumps templates with the `"MOCK"` and `"DOC"` modes. In addition to the default `"MAP"` mode, only the `"MOCK"` has an obvious application, as it can serve as an input file to a simulator code to generate mock data.

The `"DOC"` mode does not have explicit application, and was originally meant for generating automatic documentation by querying the file and rendering it on a webpage.

- __Deprecetation warning when running tests__

> There is no documentation or mention in the README.md file on how to run tests. I figured the author where using pytest, but a small statement should be included in the README.md to tell users how to run it (a simple one-liner would be appreciated).

We added the `pytest` command in the README.

> All test pass on my system, although I get a significant amounts of warning related to the usage of deprecated function. The specific warning messages I get from numpy look safe, so this is not an acceptance-blocker issue.

Which warning from `numpy` are you referring to? On my machine, I obtained a deprecation warning due to `pytest` (it is now fixed with [this commit](https://github.com/aymgal/COOLEST/commit/4211657f128f3a49bada842800b58c07cc57fa37)), and another one from `numba` (this one is related to `lenstronomy` if I am correct). On the GitHub server, only the `pytest` showed up before.

- __Other tasks__

We addressed each opened issue, and commented on them directly on [our repo](https://github.com/aymgal/COOLEST/issues).
