### 2023-07-10 4cc3dd3
First implementation of the two point correlation function (untested).
