+++ coolest/template/classes/util.py
@@ -3,10 +3,10 @@ __author__ = 'aymgal'
 # Utility functions to use throughout the API
 
 def galaxy_profile_to_id(model_type, profile_name, profile_idx, entity_idx):
-    return f'{entity_idx}-galaxy-{model_type}-{profile_idx}_{profile_name}'
+    return f'{entity_idx}-galaxy-{model_type}-{profile_idx}-{profile_name}'
 
 def mass_field_profile_to_id(profile_name, profile_idx, entity_idx):
-    return f'{entity_idx}-massfield-mass-{profile_idx}_{profile_name}'
+    return f'{entity_idx}-massfield-mass-{profile_idx}-{profile_name}'
 
 def parameter_to_id(param_name, parent_profile_id):
     return f'{parent_profile_id}-{param_name}'
