### 2023-07-17 1a0b8cc
Add methods to LensingEntityList to retrieve the list of parameter IDs with example in notebook 02
