+++ coolest/template/classes/util.py
@@ -11,23 +11,10 @@ def mass_field_profile_to_id(profile_name, profile_idx, entity_idx):
 def parameter_to_id(param_name, parent_profile_id):
     return f'{parent_profile_id}-{param_name}'
 
-def galaxy_id_to_profile(id_str):
-    galaxy_str, model_str, profile_str = id_str.split('-')
-    galaxy_idx = int(galaxy_str.split('_')[1])
-    profile_name, profile_idx = profile_str.split('_')
-    profile_idx = int(profile_idx)
-    model_type = model_str.split('_')[0]
-    return model_type, profile_name, profile_idx, galaxy_idx
+def unpack_profile_id(id_str):
+    entity_idx, _, model_type, profile_idx, profile_name = id_str.split('-')
+    return int(entity_idx), str(model_type), int(profile_idx), str(profile_name)
 
-def regul_to_id(regul_name, regul_idx):
-    return f'regularization_model-{regul_name}_{regul_idx}'
-
-def id_to_regul(id_str):
-    regul_str, model_str = id_str.split('-')
-    regul_name, regul_idx = model_str.split('_')
-    regul_idx = int(regul_idx)
-    return regul_name, regul_idx
-    
-def class_name_to_type(name):
-    import re
-    return re.sub('(?<!^)(?=[A-Z])', '_', name).lower()
+def unpack_parameter_id(id_str):
+    entity_idx, _, model_type, profile_idx, profile_name, param_name = id_str.split('-')
+    return int(entity_idx), str(model_type), int(profile_idx), str(profile_name), str(param_name)
