### 2023-07-17 8369af4
Improve error message when jsonpickle is not able to construct proper COOLEST instance
