### 2023-07-18 f109fae
Add function to retrieve parameter from ID
