### 2023-07-18 cb469d1
Fix issue when changing the template mode
