### 2023-07-19 #5567 @smsharma
# Review

- I was able to install the package by following the instructions, and successfully run the tests (modulo warnings). Perhaps there could be a more elegant way to handle the optional dependencies currently commented out in the requirements.txt file. I was also unsure of the role of "requirements_tests.txt", initially thinking this may be required to run the tests. Additionally, `ipykernel` etc could be added as an additional optional dependency in the same footing as `matplotlib` etc given the usefulness of the example notebooks.
- I found the example notebooks helpful. I was not able to fully run notebook 04, which exposes core lens model functionality -- I get the error `FileNotFoundError: [Errno 2] No such file or directory: '[root]/COOLEST/docs/notebooks/database/code_1/coolest_code_1.json'` when trying to run the initial code block for the analytic source model, and replacing the file string with `coolest_code_1_pyAPI.json` leads to further opaque errors.
- There are a few minor typos and grammatical issues in the paper e.g., "... we could built a standard based on..."; it could use another pass for proofreading and consistency. Background scientific use cases for strong gravitational lensing could also be substantiated a bit more; at the moment, the paper states "observations seem to favor different properties of dark matter and dark energy than previously thought", which I am not sure is a consensus statement.
- Very minor: The package logo does not show up correctly in GitHub dark mode (several visual elements blend in with the dark background). This could be corrected by having a mode-agnostic logo, having [context-dependent](https://github.blog/changelog/2021-11-24-specify-theme-context-for-images-in-markdown/) logos. The same goes for the main figure shown in README.md.
- More specific wording on contributions and expectations for contributing would be welcome. This could be as simple as an invitation to open an issue on GitHub for bugs or functionality requests, or a more comprehensive `CONTRIBUTING.md` file as is sometimes useful depending on context of the package.
