### 2023-07-19 dea5685
Add information about contribution to the project
