### 2023-07-19 892ca2e
Fix example template file not compatible with latest change in the template submodule
