+++ coolest/template/json.py
@@ -16,14 +16,18 @@ class JSONSerializer(object):
     """Object that write a COOLEST object to a JSON template file,
     or loads the content of a JSON template file in the COOLEST format.
 
-    For reading, either a plain JSON format or the one generated via jsonpickle can be provided.
+    For reading, either a plain JSON format or the one generated via `jsonpickle` can be provided.
     For the latter, the JSON file should end with the suffix `'_pyAPI'` (before the .json extension).
+
+    NOTE: the support for reading the template with `jsonpickle` will probably 
+    be suppressed in the future. Hence we advise users to work with the 
+    pure JSON template files instead.
     
     Parameters
     ----------
     file_path_no_ext : str
         Path to the JSON template, or the one to be created.
-        It should NOT include the .json extension nor the optional _pyAPI
+        It should NOT include the .json extension nor the optional '_pyAPI' suffix
     obj : object, optional
         Instance of the COOLEST object (from the `standard` module) 
         to be encoded, by default None
@@ -41,6 +45,9 @@ class JSONSerializer(object):
         If the provided path contains the .json extension
     """
 
+    # suffix to filename to distinguish files that can be read using jsonpickle
+    _api_suffix = '_pyAPI'
+
     def __init__(self,
                  file_path_no_ext: str, 
                  obj: object = None, 
@@ -54,8 +61,6 @@ class JSONSerializer(object):
         self._json_dir = os.path.dirname(file_path_no_ext)
         self.obj = obj
         self.indent = indent
-        # to distinguish files that can be converted back to the python API
-        self._api_suffix = '_pyAPI'
         self._check_files = check_external_files
 
     def dump_simple(self, exclude_keys=None):
@@ -85,8 +90,8 @@ class JSONSerializer(object):
 
     def load(self, verbose=False):
         """Read the JSON template file and build up the corresponding COOLEST object.
-        It will first try to load it using `jsonpickle` (if the _pyAPI file exists), 
-        otherwise it will fall back to the manual method.
+        It will first try to load the '_pyAPI' template if it exists using `jsonpickle`, 
+        otherwise it will fall back to reading the pure json template.
 
         Parameters
         ----------
@@ -101,9 +106,9 @@ class JSONSerializer(object):
         try:
             content = self.load_jsonpickle()
         except Exception as e:
-            if verbose:
-                    print(f"Failed reading with jsonpickle, trying reading pure json"
-                          f" (original error: {e})")
+            if verbose is True:
+                print(f"Failed reading '{self._api_suffix}' template with jsonpickle, "
+                    f"now trying with the pure json template (original error: {e})")
             content = self.load_simple()
         assert isinstance(content, COOLEST)
         return content
@@ -122,6 +127,8 @@ class JSONSerializer(object):
             COOLEST object that corresponds to the JSON template
         """
         json_path = self.path + '.json'
+        if not os.path.exists(json_path):
+            raise ValueError(f"Template file at '{json_path}' does not exist.")
         with open(json_path, 'r') as f:
             content = json.loads(f.read())
         if not as_object:
@@ -138,6 +145,8 @@ class JSONSerializer(object):
             COOLEST object that corresponds to the JSON template
         """
         json_path = self.path + self._api_suffix + '.json'
+        if not os.path.exists(json_path):
+            raise ValueError(f"Template file at '{json_path}' does not exist.")
         with open(json_path, 'r') as f:
             content = jsonpickle.decode(f.read())
         return content  # COOLEST object
