### 2023-07-20 f6d9ff0
Add function to get list of parameter instances
