### 2023-07-20 3a3c814
Improve handling of coordinates in two_point_correlation
