### 2023-07-20 1e2de09
Add (default) option to use coordinates from pixelated profile in 2-point correlation
