+++ docs/notebooks/database/vkl_paper_3.3_dummy/fff_gauss_gauss_n3/output/mn-summary.txt
@@ -1,0 +1,2 @@
+    0.341259587496498207E+00    0.769265173237393407E+00    0.472775762437915970E+00    0.329451733313615214E+00    0.391271707377444489E+00    0.282338413416573009E+00    0.586573314927700595E+00    0.845784666216911729E+00    0.363222303031605409E+00    0.907173619638450113E+00    0.699016046238694955E+00    0.312251498343505754E-01    0.121230122772318455E+00    0.247571748225176198E-01    0.341367452534044810E-01    0.484660360500433832E-01    0.152580389373966133E-01    0.141342246112229133E-01    0.659760937256865122E-02    0.981099923193630430E-02    0.962778121310256876E-02    0.303079799854849806E-01    0.350791831652335073E+00    0.666190790930716203E+00    0.479002414410515387E+00    0.340163597540492024E+00    0.422943959987497675E+00    0.267496495077767127E+00    0.573101963852661545E+00    0.848395011529152088E+00    0.357018023554120711E+00    0.900263643212385500E+00    0.709277096186469280E+00    0.350791831652335073E+00    0.666190790930716203E+00    0.479002414410515387E+00    0.340163597540492024E+00    0.422943959987497675E+00    0.267496495077767127E+00    0.573101963852661545E+00    0.848395011529152088E+00    0.357018023554120711E+00    0.900263643212385500E+00    0.709277096186469280E+00    0.180826870439330523E+05    0.181211735960695441E+05
+    0.341259587495850447E+00    0.769265173235937350E+00    0.472775762437020630E+00    0.329451733312991490E+00    0.391271707376702804E+00    0.282338413416038381E+00    0.586573314926587486E+00    0.845784666215311787E+00    0.363222303030917792E+00    0.907173619636730266E+00    0.699016046237368571E+00    0.312251498378661139E-01    0.121230122776820062E+00    0.247571748310391956E-01    0.341367452563803228E-01    0.484660360529921147E-01    0.152580389423263175E-01    0.141342246343554648E-01    0.659760947498176653E-02    0.981099924464711934E-02    0.962778129413408533E-02    0.303079800007776644E-01    0.350791831652335073E+00    0.666190790930716203E+00    0.479002414410515387E+00    0.340163597540492024E+00    0.422943959987497675E+00    0.267496495077767127E+00    0.573101963852661545E+00    0.848395011529152088E+00    0.357018023554120711E+00    0.900263643212385500E+00    0.709277096186469280E+00    0.350791831652335073E+00    0.666190790930716203E+00    0.479002414410515387E+00    0.340163597540492024E+00    0.422943959987497675E+00    0.267496495077767127E+00    0.573101963852661545E+00    0.848395011529152088E+00    0.357018023554120711E+00    0.900263643212385500E+00    0.709277096186469280E+00    0.180827731187208592E+05    0.181211735960695441E+05
