+++ joss/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 
 # Summary
 
-Any mass concentration in the Universe, luminous or dark, from vast galaxy clusters to stars within galaxies, can be studied through its gravitationally induced deflections of light rays from background sources – a phenomenon known as Strong Gravitational Lensing (SGL). This has several cutting-edge applications, for example: measuring the Hubble constant and shedding more light into the apparent tension between early and late Universe, detecting the presence of massive subhalos within distant galaxies that can constrain different dark matter models, and studying a galaxy’s mass partition between baryons and dark matter with direct implications on galaxy evolution.
+Any mass concentration in the Universe, luminous or dark, from vast galaxy clusters to stars within galaxies, can be studied through its gravitational deflection of light rays from background sources. This phenomenon, in its most impressive regime, is known as Strong Gravitational Lensing (SGL). It has several cutting-edge applications, for example: measuring the Hubble constant and shedding more light into the apparent tension between early and late Universe, detecting the presence of massive subhalos within distant galaxies that can constrain different dark matter models, and studying a galaxy’s mass partition between baryons and dark matter with direct implications on galaxy evolution.
 
 Extracting information from SGL data requires the careful analysis of images of gravitational lenses, a process referred to as _lens modeling_, in order to generate an image of the lens based on models of mass and light distributions of the different physical objects in play (e.g., galaxies, quasars). In this paper we call a _lens model_ the full set of model components, including all mass and light models as well as the point spread function (PSF) model. Over the past twenty years, several lens modeling codes have been developed and used in published works. Unfortunately, there is currently no efficient and systematic way to access these published results and use them directly for new studies, which slows down new research and causes a waste of research time. The reason is simple: these modeling codes being based on different methods and conventions, bridging the gap between them is a challenging task.
 
