### 2023-07-21 d775d62
Removed several files associated with the notebook examples that were not really required.
