+++ docs/notebooks/database/vkl_paper_3.4_dummy/fff_exp_gauss_n3/output/mn-summary.txt
@@ -1,2 +1,0 @@
-    0.382555208943496206E+00    0.701123545335680554E+00    0.407123008366949868E+00    0.241009938951896124E+00    0.405772707604553362E+00    0.232245931152529672E+00    0.629139980324069925E+00    0.839595839155829138E+00    0.418732075432536432E+00    0.830832506690910799E+00    0.610217565530671857E+00    0.138144554015610203E-01    0.555871228862392860E-01    0.993326765563057844E-02    0.139621672866222062E-01    0.185481430439071796E-01    0.313881398935499578E-02    0.438181525048563728E-02    0.138480384117767759E-01    0.779027701443212317E-01    0.146329698952415918E-01    0.642500069309036026E-01    0.377283707389349354E+00    0.678492293418291981E+00    0.404869725126593638E+00    0.237498893190466270E+00    0.413916067044635350E+00    0.231953905483584022E+00    0.627703444421810031E+00    0.837864725950277167E+00    0.409821561874827156E+00    0.827593951833909491E+00    0.631586718535200986E+00    0.377283707389349354E+00    0.678492293418291981E+00    0.404869725126593638E+00    0.237498893190466270E+00    0.413916067044635350E+00    0.231953905483584022E+00    0.627703444421810031E+00    0.837864725950277167E+00    0.409821561874827156E+00    0.827593951833909491E+00    0.631586718535200986E+00    0.203425059458397664E+05    0.203878746918231664E+05
-    0.382555208949171777E+00    0.701123545346085231E+00    0.407123008372990425E+00    0.241009938955472597E+00    0.405772707610574657E+00    0.232245931155975888E+00    0.629139980333405124E+00    0.839595839168289948E+00    0.418732075438748463E+00    0.830832506703239382E+00    0.610217565539726614E+00    0.138144553230919211E-01    0.555871228210250354E-01    0.993326753192837544E-02    0.139621672558585055E-01    0.185481429781857235E-01    0.313881386187648756E-02    0.438181458031409576E-02    0.138480380340751021E-01    0.779027701282089535E-01    0.146329695453836679E-01    0.642500068883846559E-01    0.377283707389349354E+00    0.678492293418291981E+00    0.404869725126593638E+00    0.237498893190466270E+00    0.413916067044635350E+00    0.231953905483584022E+00    0.627703444421810031E+00    0.837864725950277167E+00    0.409821561874827156E+00    0.827593951833909491E+00    0.631586718535200986E+00    0.377283707389349354E+00    0.678492293418291981E+00    0.404869725126593638E+00    0.237498893190466270E+00    0.413916067044635350E+00    0.231953905483584022E+00    0.627703444421810031E+00    0.837864725950277167E+00    0.409821561874827156E+00    0.827593951833909491E+00    0.631586718535200986E+00    0.203425671331522681E+05    0.203878746918231664E+05
