### 2023-07-21 84dee9c
Minor edit of the README
