### 2023-07-21 #5567 @aymgal
Hi @smsharma , thank you for your comments! Here are the answers and details on where to find the changes in the code, doc and the paper.

- __Content of the in paper__

The introduction of the paper has been updated to address your comment. The changes can be seen [this PR diff](https://github.com/aymgal/COOLEST/pull/47/files).

- __Unclear test requirements and optional dependencies__

The `requirements_tests.txt` is indeed destined to run the unit tests with `pytest`. We made that clearer in the README. We also made some improvements in the `setup.py` such that running the command `pip install ".[opt]"` will automatically installs optional dependencies as well (see [this PR](https://github.com/aymgal/COOLEST/pull/45)).

- __Information regarding external contributions__

Thank you for suggesting this, it's a useful thing to have indeed! We added more information with [this commit](https://github.com/aymgal/COOLEST/commit/dea5685e9088a06bff5b4a93c27046ef814bdc93) to encourage contributions by the lensing community in a CONTRIBUTION.md document.

- __Problem with example notebook__

This has been fixed with [this commit](https://github.com/aymgal/COOLEST/commit/892ca2ed8836535d586a62a69adf22011eb6be29), and the error message should now be clearer. The issue was that one of the example coolest template file (from 'code_1') contained fields that do no more corresponds to actual Python attributes in the `coolest.template` hierarchy, which caused `jsonpickle` to fail constructing Python objects.

- __Having context-dependent logos__

Thank you for the suggestion: this is now fixed with [this PR](https://github.com/aymgal/COOLEST/pull/44).

Please let us know if something more should be addressed.
