### 2023-07-22 #5567 @AlexandreAdam
# Review #2
## List of conventions
> If possible, I would prefer to have the list of conventions in the documentation / GitHub only, such that it can be completed and improved along its adoption by researchers and the feedback from the community. Although the current stated conventions should not change, it is still an eventuality and it is probable that other conventions get added to the list in the future. Moreover, as the paper is already over the "standard" size, this would make it significantly longer.

That works with me. The conventions are clearly stated and easily accessible from the documentation or the README.md.

## Tests of the effective quantities computation
> We added unit test for various input parameters and grid resolution in test/test. For these tests, we quote a maximum relative error of 5% on Einstein radius, effective kappa slope and effective half-light radius computations (the error being most of the time inferior to 5%).
>We note that COOLEST is not meant to reproduce the original modeling results (this would be unrealistic), but to provide tools to analyse lens models uniformly, independently from the original modeling code as much as possible. Therefore, differences in the values of effective quantities compared to original lens modeling codes and theoretical values are expected, but does not impact the goal of the standard.

Thank you for the clarification.

## Clarification of the concept of lensing entities

The new formulation in the documention is much more clear and to the point.

## Clarification on the template "modes"

The documentation is now very clear on the `modes` keyword.

## Deprecetation warning when running tests

The specific DeprecationWarning I get is the following

```
DeprecationWarning: `np.long` is a deprecated alias for `np.compat.long`. To silence this warning, use `np.compat.long` by itself. In the likely event your code does not need to work on Python 2 you can use the builtin `int` for which `np.compat.long` is itself an alias. Doing this will not modify any behaviour and is safe. When replacing `np.long`, you may wish to use e.g. `np.int64` or `np.int32` to specify the precision. If you wish to review your current use, check the release note link for additional information.
  Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
```

## Issues

All issues related to my previous review were resolved satisfactorily.

## Conclusion

In conclusion, I am pleased to report that the code has met all the criterias in the checklist to my satisfaction.

Throughout the assessment process, it was clear to me that the authors have dedicated significant efforts to ensure the quality and robustness of the code. The code is modular, well tested and well documented. Its purpose is clear and timely.

In light of these observations, I commend the authors for their meticulous attention to details. It is my conclusion that the submitted code aligns with the JOSS standards of a scientific open source software.

I, therefore, recommend this code for a JOSS publication without reservation. Its adoption would be a valuable contribution to the gravitational lensing community.
