### 2023-07-23 02f0ab8
Remove redundant code
