### 2023-07-24 3aef492
Using the .latex_str of each parameter as the default label.
