### 2023-07-29 #5567 @smsharma
Thanks very much @aymgal for addressing all of the comments, in particular the comprehensive contribution guidelines. At this point I can enthusiastically recommend acceptance in JOSS, and believe this will be a great tool to aid in gravitational lens modeling in a standardized way.
