### 2023-07-30 #5567 @aymgal
Thanks a lot @AlexandreAdam and @smsharma for your detailed and serious reviews, as well as your suggestions to improve the standard we propose. We really appreciate it. Happy to see this going through the JOSS process!
