### 2023-07-31 1b306dd
Add ability to remove colorbar in plotting routines
