+++ coolest/api/plotting.py
@@ -53,25 +53,28 @@ class ModelPlotter(object):
         #cmap_colors[0,:] = [0.15, 0.15, 0.15, 1.0]  # Set the color of the very first value to gray
         #self.cmap_flux_mod = ListedColormap(cmap_colors)
 
-    def plot_data_image(self, ax, title=None, norm=None, cmap=None, neg_values_as_bad=False):
+    def plot_data_image(self, ax, title=None, norm=None, cmap=None, 
+                        neg_values_as_bad=False, add_colorbar=True):
         """plt.imshow panel with the data image"""
         if cmap is None:
             cmap = self.cmap_flux
         coordinates = util.get_coordinates(self.coolest)
         extent = coordinates.plt_extent
         image = self.coolest.observation.pixels.get_pixels(directory=self._directory)
-        ax, cb = self._plot_regular_grid(ax, image, extent=extent, 
+        ax, im = self._plot_regular_grid(ax, image, extent=extent, 
                                 cmap=cmap,
                                 neg_values_as_bad=neg_values_as_bad, 
                                 norm=norm)
-        cb.set_label("flux")
+        if add_colorbar:
+            cb = plut.nice_colorbar(im, ax=ax, max_nbins=4)
+            cb.set_label("flux")
         if title is not None:
             ax.set_title(title)
         return image
 
     def plot_surface_brightness(self, ax, title=None, coordinates=None, 
                                 extent=None, norm=None, cmap=None, neg_values_as_bad=True,
-                                plot_points_irreg=False, kwargs_light=None):
+                                plot_points_irreg=False, add_colorbar=True, kwargs_light=None):
         """plt.imshow panel showing the surface brightness of the (unlensed)
         lensing entity selected via kwargs_light (see ComposableLightModel docstring)"""
         if kwargs_light is None:
@@ -83,7 +86,7 @@ class ModelPlotter(object):
             x, y = coordinates.pixel_coordinates
             image = light_model.evaluate_surface_brightness(x, y)
             extent = coordinates.plt_extent
-            ax, cb = self._plot_regular_grid(ax, image, extent=extent, cmap=cmap,
+            ax, im = self._plot_regular_grid(ax, image, extent=extent, cmap=cmap,
                                              neg_values_as_bad=neg_values_as_bad, 
                                              norm=norm)
         else:
@@ -92,24 +95,26 @@ class ModelPlotter(object):
                 extent = extent_model
             if isinstance(values, np.ndarray) and len(values.shape) == 2:
                 image = values
-                ax, cb = self._plot_regular_grid(ax, image, extent=extent, 
+                ax, im = self._plot_regular_grid(ax, image, extent=extent, 
                                         cmap=cmap, 
                                         neg_values_as_bad=neg_values_as_bad,
                                         norm=norm)
             else:
                 points = values
-                ax, cb = self._plot_irregular_grid(ax, points, extent, norm=norm, cmap=cmap, 
+                ax, im = self._plot_irregular_grid(ax, points, extent, norm=norm, cmap=cmap, 
                                                    neg_values_as_bad=neg_values_as_bad,
                                                    plot_points=plot_points_irreg)
                 image = None
-        cb.set_label("flux")
+        if add_colorbar:
+            cb = plut.nice_colorbar(im, ax=ax, max_nbins=4)
+            cb.set_label("flux")
         if title is not None:
             ax.set_title(title)
         return image, coordinates
 
     def plot_model_image(self, ax, supersampling=5, convolved=False, title=None,
[…]
