### 2023-07-31 b10ab1c
Add option to set the global title for MultiPlotter
