### 2023-07-31 3ace975
Passing colors to the 1D plots in corner plot.
