### 2023-08-01 #5567 @plaplant
@AlexandreAdam @smsharma thank you both for your very thorough reviews! Happy to hear that this paper is ready to go.

@aymgal both reviewers have recommended acceptance of the submission, so we can move forward with the final acceptance checks. If the software has changed in the course of the review (excluding changes to the paper/manuscript itself), please make a new tagged release of the repository. After that, please archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or some other long-term hosting platform). Once those are done, please post the version number and DOI of the archive in this thread. Let me know if you have any questions!
