### 2023-08-01 ca5f6b2
Fix 1D hist color in corner plotter and other minor changes
