### 2023-08-01 b1f1772
Change error message when reading with jsonpickle
