### 2023-08-01 edb2894
Minor editions to JOSS paper
