+++ joss/paper.md
@@ -59,7 +59,7 @@ In SGL studies, the lens modeling step is often the most time-consuming. The com
 
 Such a variety of data sets and scientific objectives have led to the development of different lens modeling codes. These codes may be written in different programming languages and generally based on fundamentally different assumptions, some are not open-source, and some may not be well-documented. Consequently, when a new SGL study gets published, it is very challenging and time-consuming (sometimes impossible) to use these new results in any subsequent analysis, should it be with the same code or with a different one that is better suited to the new objective. Moreover, comparing lens models with previously published ones is as challenging, exactly for the same reasons. In the past, only a few studies have tried to compare a selection of output quantities from different lens modeling analyses: in the context of lensing by galaxy clusters [@refsdal2016; @Meneghetti2017], or for time-delay cosmography with lensed quasars [@tdlmc2021; @tdcosmo2022]. However, until now, there has been no standard way to describe, store and share lens modeling products.
 
-This is the motivation behind COOLEST: because lens modeling products follow the same theoretical foundations, we could built a standard based on precise conventions so that lens models can be described _independently of the original modeling code_. Important lens modeling products typically include the lens mass distribution, the unconvolved surface brightness of both the lens and the (unlensed) source galaxies, as well as a model of the point spread function of the instrument. COOLEST offers a simple and human-readable way to describe a lens model, summarized in a single JSON template file, and optionally links to external files (typically in the standard astronomical [FITS](https://heasarc.gsfc.nasa.gov/docs/heasarc/fits.html) format for images and tables, or in the [`pickle`](https://docs.python.org/3/library/pickle.html) format for high-dimensional arrays), all stored within a single directory. At the core of the template is a list of _lensing entities_, a new concept which allows researchers to describe the gravitational lens directly in terms of physical objects (e.g., galaxies or quasars), which is more intuitive than the abstract description used within modeling codes. It also enables a novel way to cross-reference physical objects among different analyses, including non-lensing analyses such as those focusing on galaxy evolution.
+This is the motivation behind COOLEST: because lens modeling products follow the same theoretical foundations, we were able to build a standard based on a set of specific conventions so that lens models can be described _independently of the original modeling code_. Important lens modeling products typically include the lens mass distribution, the unconvolved surface brightness of both the lens and the (unlensed) source galaxies, as well as a model of the point spread function of the instrument. COOLEST offers a simple and human-readable way to describe a lens model, summarized in a single JSON template file, and optionally links to external files (typically in the standard astronomical [FITS](https://heasarc.gsfc.nasa.gov/docs/heasarc/fits.html) format for images and tables, or in the [`pickle`](https://docs.python.org/3/library/pickle.html) format for high-dimensional arrays), all stored within a single directory. At the core of the template is a list of _lensing entities_, a new concept which allows researchers to describe the gravitational lens directly in terms of physical objects (e.g., galaxies or quasars), which is more intuitive than the abstract description used within modeling codes. It also enables a novel way to cross-reference physical objects among different analyses, including non-lensing analyses such as those focusing on galaxy evolution.
 
 In summary, COOLEST aims at bridging the existing gap between independent SGL analyses and science goals, by providing a standardized way of describing and sharing lens models, within and outside the lensing community.
 
