### 2023-08-02 601359f
Improve loading of json template
