### 2023-08-02 4dccd4e
Only read the necessary columns are now read from the chain file (prevent columns that contain e.g. arrays to be loaded unnecessarily)
