### 2023-08-02 #5567 @aymgal
@plaplant Thanks! Here is the [zenodo upload,](https://zenodo.org/record/8207512) and the corresponding DOI is 10.5281/zenodo.7900798. I created a tag `v0.1.0` on the repo, and merged the `joss` branch to the `main`.
