### 2023-08-02 b1ab89a
Update example notebooks
