+++ README.md
@@ -8,7 +8,7 @@
 [![Documentation Status](https://readthedocs.org/projects/coolest/badge/?version=latest)](https://coolest.readthedocs.io/en/latest/?badge=latest)
 [![Tests](https://github.com/aymgal/COOLEST/actions/workflows/python-package.yml/badge.svg?branch=main)](https://github.com/aymgal/COOLEST/actions/workflows/python-package.yml)
 [![Coverage Status](https://coveralls.io/repos/github/aymgal/COOLEST/badge.svg)](https://coveralls.io/github/aymgal/COOLEST)
-
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8207512.svg)](https://doi.org/10.5281/zenodo.8207512)
 
 ## What is COOLEST?
 
