+++ docs/notebooks/index.md
@@ -1,6 +1,6 @@
 # COOLEST example notebooks
 
-We provide a series of [Python notebooks](https://github.com/aymgal/COOLEST/tree/master/docs/notebooks) for showing different use cases of the COOLEST interface.
+We provide a series of [Python notebooks](https://github.com/aymgal/COOLEST/tree/main/docs/notebooks) for showing different use cases of the COOLEST interface.
 
 ```{toctree}
 ---
@@ -12,5 +12,6 @@ List supported options <01-list_supported_options.ipynb>
 Generate a JSON template file <02-generate_template.ipynb>
 Update a template in 2 ways <03-load_and_update_template.ipynb>
 Plot models side-to-side <04-plot_lens_models.ipynb>
-Analysis of surface brightness models <05-two_point_correlation.ipynb>
+Surface brightness analysis <05-two_point_correlation.ipynb>
+Corner plot for posteriors <06-make_corner_plot.ipynb>
 ```
