### 2023-08-02 d4147e2
Added QLens to the list of compliant lens modeling softwares
