+++ docs/examples/existing_interfaces.md
@@ -10,4 +10,8 @@ The following modeling codes already have an interface with COOLEST:
 
 - [__*VKL*__](https://github.com/gvernard/verykool) ([Vernardos et al. 2022](https://ui.adsabs.harvard.edu/abs/2022MNRAS.516.1347V/abstract)).
 
-If you have questions or comments about the development of a COOLEST interface for your modeling code, please [reach out to us](mailto:aymeric.galan@gmail.com). In case of an issue with the COOLEST API, we advise you to open an [issue on GitHub](https://github.com/aymgal/COOLEST/issues/new).
+- [__*MOLET*__](https://github.com/gvernard/molet) ([Vernardos et al. 2021](https://ui.adsabs.harvard.edu/abs/2022MNRAS.511.4417V/abstract)).
+
+- __*QLens*__ (Minor et al., in prep.).
+
+If you have questions or comments about the development of a COOLEST interface for your modeling code, please [reach out to us](mailto:aymeric.galan@gmail.com). In case of an issue with the COOLEST API, we recommend you to careful read [this document](https://github.com/aymgal/COOLEST/blob/main/CONTRIBUTING.md) and eventually [open an issue](https://github.com/aymgal/COOLEST/issues/new) on GitHub.
