### 2023-08-02 #5567 @aymgal
We just added some minor changes to the paper.md, the very last version is on the main branch.
