### 2023-08-03 #5567 @aymgal
Hi @plaplant it looks like the latest compiled paper is not the one on the `main` branch (which is now the latest up to date). Should I update the `joss` branch as well? Or is it possible to ensure the bot compiles the paper from the `main` branch?
