### 2023-08-03 0eefb79
Fixing journal name macros
