+++ joss/paper.bib
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{herculens2022,
        author = {{Galan}, A. and {Vernardos}, G. and {Peel}, A. and {Courbin}, F. and {Starck}, J. -L.},
         title = "{Using wavelets to capture deviations from smoothness in galaxy-scale strong lenses}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {galaxies: structure, dark matter, methods: data analysis, gravitation, gravitational lensing: strong, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = dec,
@@ -38,7 +38,7 @@ archivePrefix = {arXiv},
 @ARTICLE{vkl2022,
        author = {{Vernardos}, G. and {Koopmans}, L.~V.~E.},
         title = "{The very knotty lenser: Exploring the role of regularization in source and potential reconstructions using Gaussian process regression}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = oct,
@@ -56,7 +56,7 @@ archivePrefix = {arXiv},
 @ARTICLE{molet2022,
        author = {{Vernardos}, G.},
         title = "{Simulating time-varying strong lenses}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, gravitational lensing: micro, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2022,
         month = apr,
@@ -87,7 +87,7 @@ archivePrefix = {arXiv},
 @ARTICLE{gigalens2022,
        author = {{Gu}, A. and {Huang}, X. and {Sheu}, W. and {Aldering}, G. and {Bolton}, A.~S. and {Boone}, K. and {Dey}, A. and {Filipp}, A. and {Jullo}, E. and {Perlmutter}, S. and {Rubin}, D. and {Schlafly}, E.~F. and {Schlegel}, D.~J. and {Shu}, Y. and {Suyu}, S.~H.},
         title = "{GIGA-Lens: Fast Bayesian Inference for Strong Gravitational Lens Modeling}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {GPU computing, Markov chain Monte Carlo, Strong gravitational lensing, Einstein rings, Bayesian statistics, Cosmology, 1969, 1889, 1643, 451, 1900, 343, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Statistics - Applications},
          year = 2022,
         month = aug,
@@ -106,7 +106,7 @@ archivePrefix = {arXiv},
 @ARTICLE{glee2010,
        author = {{Suyu}, S.~H. and {Halkola}, A.},
         title = "{The halos of satellite galaxies: the companion of the massive elliptical lens SL2S J08544-0121}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
[…]
