+++ coolest/api/plotting.py
@@ -515,6 +515,8 @@ def plot_corner(parameter_id_list,chain_objs,chain_dirs,chain_names=None,point_e
                     values.append(val)
             point_estimates.append(values)
 
+
+            
     mcsamples = []
     for i in range(0,len(chain_objs)):
         chain_file = os.path.join(chain_dirs[i],chain_objs[i].meta["chain_file_name"]) # Here get the chain file path for each coolest object
@@ -579,6 +581,7 @@ def plot_corner(parameter_id_list,chain_objs,chain_dirs,chain_names=None,point_e
                         line_args=[{'ls':'-', 'lw': 2, 'color': c} for c in colors], 
                         contour_colors=colors)
 
+
     my_linestyles = ['solid','dotted','dashed','dashdot']
     my_markers    = ['s','^','o','star']
 
@@ -602,5 +605,25 @@ def plot_corner(parameter_id_list,chain_objs,chain_dirs,chain_names=None,point_e
                 else:
                     pass    
 
-                
+
+    # Set default ranges for angles
+    for i in range(0,len(parameter_id_list)):
+        dum = parameter_id_list[i].split('-')
+        name = dum[-1]
+        if name in ['phi','phi_ext']:
+            xlim = image.subplots[index,index].get_xlim()
+            #print(xlim)
+        
+            if xlim[0] < -90:
+                for ax in image.subplots[index:,index]:
+                    ax.set_xlim(left=-90)
+                for ax in image.subplots[index,:index]:
+                    ax.set_ylim(bottom=-90)
+            if xlim[1] > 90:
+                for ax in image.subplots[index:,index]:
+                    ax.set_xlim(right=90)
+                for ax in image.subplots[index,:index]:
+                    ax.set_ylim(top=90)
+
+            
     return image
